Move shouldRevealHiddenInfo into its own lib/hidden_info module
abstract_game.ts is about the contract every game implements; the
hidden-info reveal rule applies only to the handful of variants with
per-player visibility. Keep the abstract file focused on the base
contract and give the helper a home that scales if we add more
hidden-info utilities (e.g. for the observer-as-player follow-up).
